third_party/WebKit/LayoutTests/fast/events/uievent-with-inputdevice-expected.txt

This tests that UIEvent will have sourceDevice set to be null by default, and it can also be passed when initialization.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event is non-null.
PASS event.sourceDevice is null
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
PASS successfullyParsed is true

TEST COMPLETE
